Update rat.md
Verified all links. Reported one offline link. I suggest to add a "Description" column for inserting the scope of the website. For the offline links, I suggest to remove those rows or find for new links (if they are online on different links). I suggest to remove them and allow community to fill this table with the most updated links. I suggest to do a double check for offline links.